Added two images in miscellaneous for header background.
Added a different image to the slideshow.
---

Changed border radius for the image.
---

Added unique captions for each image in the html and corresponding css.

Changed the id of the li elements in the ul(carousel-dots) from "carousel" to "dots".

---

Added functionality in the scripts.js to dynamically change the caption along with the images.