Added a display photo for the about me section.
The current state gives a good idea for a background trick.
Making the display photo in the about me section the same image as the header makes it seem like the image in the header is continuously scrolling. A good way to add an effect.

Other changes were to formatting.